requirements.txt
Flask
requests
Flask-PyMongo
Flask-Mail
python-dotenv
epayco
